[#8] Move KafkaConsumerConfigProperties to package "consumer".
