+++ cli/src/main/java/org/eclipse/hono/cli/app/KafkaReceiver.java
@@ -23,7 +23,7 @@ import javax.annotation.PostConstruct;
 
 import org.eclipse.hono.cli.AbstractCliClient;
 import org.eclipse.hono.kafka.client.HonoTopic;
-import org.eclipse.hono.kafka.client.KafkaConsumerConfigProperties;
+import org.eclipse.hono.kafka.client.consumer.KafkaConsumerConfigProperties;
 import org.eclipse.hono.util.EventConstants;
 import org.eclipse.hono.util.MessageHelper;
 import org.eclipse.hono.util.Strings;
