+++ clients/kafka-common/src/main/java/org/eclipse/hono/kafka/client/consumer/KafkaConsumerConfigProperties.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.kafka.client;
+package org.eclipse.hono.kafka.client.consumer;
 
 import java.util.Collections;
 import java.util.HashMap;
